@KyleAMathews 2021-05-05
Good deleted/added code ratio 👍 

Looks like a ~12% speedup for 1000 queries (using my handy one-liner — `split(`\n`).map(l => parseFloat(l.split(` `).slice(-1)[0].split(`/`)[0])).reduce((a,b) => a + b) /  9`)

I [got 20%](https://github.com/gatsbyjs/gatsby/commit/78d6986069354ed2904ebbcdbd05cd42a3f26ce0) but that was with 100k queries — the more queries the bigger the speedup I'd assume given better-queue's sorting logic would get slower on bigger datasets.