@vladar 2021-05-05
Yeah, also benchmark setup is probably different. I believe my benchmark still does a pass of fast filters (since I am not filtering on the `id` field). Filtering by `id` has less overhead and so the perf benefit is more noticeable.